ai-search-engine-api/requirements.txt
docstring_inheritance  
click  
curl_cffi  
nest-asyncio  
g4f>=0.2.2.3  
g4f[webdriver]  
rich  
beautifulsoup4
markdownify
python-dotenv
requests  
google-generativeai  
lxml>=5.2.2
termcolor  
orjson  
PyYAML  
tls_client  
clipman  
playsound  
ollama  
pyfiglet  
pillow  
requests_html  
bson  
cloudscraper  
emoji  
openai  
prompt-toolkit  
fake_useragent  
primp
beautifulsoup4==4.12.3
blessed==1.20.0
Brotli==1.1.0
bs4==0.0.1
requests
soupsieve==2.5
orjson
spinners==0.0.24
sse-starlette==2.1.2
stack-data==0.6.3
starlette==0.37.2
StrEnum==0.4.15
style==1.1.0
termcolor==2.4.0
tiktoken==0.6.0
tldextract==5.1.2
tls-client==1.0.1
tokenizers==0.19.1
tokentrim==0.1.13
toml==0.10.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
fastapi
pyreqwest_impersonate
aiohttp
mechanize
easygoogletranslate
pillow
primp